pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "jobbot"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "requests>=2.31.0",
  "beautifulsoup4>=4.12.0",
  "PyYAML>=6.0.0",
  "typer>=0.9.0",
]

[project.scripts]
jobbot = "jobbot.app.cli:main"

[tool.setuptools]
packages = ["jobbot"]
